requirements.txt
python-dotenv
requests
python-telegram-bot

